+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/AppLifecycleProvider.kt
@@ -1,0 +1,7 @@
+package au.com.shiftyjelly.pocketcasts.repositories.playback
+
+import kotlinx.coroutines.flow.StateFlow
+
+interface AppLifecycleProvider {
+    val isInForeground: StateFlow<Boolean>
+}