+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 8.10
------
 
 
 8.9
 -----
+*   New Features
+    *   Display playback errors on bottom bar
+        ([#5188](https://github.com/Automattic/pocket-casts-android/pull/5188))
 *   Updates
     *   Show progress while mini player is buffering
         ([#5165](https://github.com/Automattic/pocket-casts-android/pull/5165))